Remove overlay functionality
We have multiple ways of generating docs, and including docs overlays
for the Docker:Image resource was done via this tool. We have since
moved to keeping "extra" documentation as part of the provider
repository, examples include https://github.com/pulumi/pulumi-databricks/pull/38/
and https://github.com/pulumi/pulumi-kubernetes/pull/1999, as well as
altering documentation in our terraform provider forks.

We have the ability to keep example documentation in the schema for
Docker:Image as of pulumi-docker v4, and therefore this functionality is
obsolete, and removal is a key step in consolidating how we do
supplemental provider documentation.